Moved vesting action out of widget package.
