+++ packages/stateful/actions/core/treasury/ManageVesting/BeginVesting.stories.tsx
@@ -7,9 +7,9 @@ import {
   makeDaoProvidersDecorator,
   makeReactHookFormDecorator,
 } from '@dao-dao/storybook'
-import { TokenType } from '@dao-dao/types'
+import { TokenType, VestingPaymentsWidgetVersion } from '@dao-dao/types'
 
-import { VestingPaymentsWidgetVersion } from '../../types'
+import { EntityDisplay } from '../../../../components'
 import { BeginVesting } from './BeginVesting'
 
 export default {
@@ -67,6 +67,7 @@ Default.args = {
     ],
     vestingFactoryOwner: { loading: false, data: undefined },
     AddressInput,
+    EntityDisplay,
     createCw1WhitelistOwners: () => alert('createCw1WhitelistOwners'),
     creatingCw1WhitelistOwners: false,
   },
