+++ packages/stateful/components/vesting/index.ts
@@ -1,0 +1,2 @@
+export * from './VestingStakingModal'
+export * from './VestingPaymentCard'
