+++ packages/stateful/recoil/selectors/index.ts
@@ -5,4 +5,5 @@ export * from './nft'
 export * from './profile'
 export * from './token'
 export * from './treasury'
+export * from './vesting'
 export * from './wallet'
