+++ packages/stateful/widgets/react/index.ts
@@ -1,1 +1,2 @@
+export * from './useWidget'
 export * from './useWidgets'
