--- packages/stateful/widgets/widgets/Press/constants.ts
@@ -1,1 +1,0 @@
-export const PRESS_WIDGET_ID = 'press'
