+++ packages/stateful/widgets/widgets/Press/index.ts
@@ -3,6 +3,7 @@ import { ArticleOutlined } from '@mui/icons-material'
 import {
   ActionCategoryKey,
   Widget,
+  WidgetId,
   WidgetLocation,
   WidgetVisibilityContext,
 } from '@dao-dao/types'
@@ -10,13 +11,12 @@ import {
 import { makeCreatePostActionMaker } from './actions/CreatePost'
 import { makeDeletePostActionMaker } from './actions/DeletePost'
 import { makeUpdatePostActionMaker } from './actions/UpdatePost'
-import { PRESS_WIDGET_ID } from './constants'
 import { PressEditor as Editor } from './PressEditor'
 import { Renderer } from './Renderer'
 import { PressData } from './types'
 
 export const PressWidget: Widget<PressData> = {
-  id: PRESS_WIDGET_ID,
+  id: WidgetId.Press,
   Icon: ArticleOutlined,
   location: WidgetLocation.Tab,
   visibilityContext: WidgetVisibilityContext.Always,
