--- packages/stateful/widgets/widgets/RetroactiveCompensation/constants.ts
@@ -1,1 +1,0 @@
-export const RETROACTIVE_COMPENSATION_WIDGET_ID = 'retroactive'
