+++ packages/stateful/widgets/widgets/RetroactiveCompensation/index.ts
@@ -1,12 +1,16 @@
 import { PaidOutlined } from '@mui/icons-material'
 
-import { Widget, WidgetLocation, WidgetVisibilityContext } from '@dao-dao/types'
+import {
+  Widget,
+  WidgetId,
+  WidgetLocation,
+  WidgetVisibilityContext,
+} from '@dao-dao/types'
 
-import { RETROACTIVE_COMPENSATION_WIDGET_ID } from './constants'
 import { Renderer } from './Renderer'
 
 export const RetroactiveCompensationWidget: Widget = {
-  id: RETROACTIVE_COMPENSATION_WIDGET_ID,
+  id: WidgetId.RetroactiveCompensation,
   Icon: PaidOutlined,
   location: WidgetLocation.Tab,
   visibilityContext: WidgetVisibilityContext.Always,
