--- packages/stateful/widgets/widgets/VestingPayments/constants.ts
@@ -1,1 +1,0 @@
-export const VESTING_PAYMENTS_WIDGET_ID = 'vesting'
