+++ packages/stateful/widgets/widgets/VestingPayments/index.ts
@@ -1,20 +1,19 @@
 import { PaidOutlined } from '@mui/icons-material'
 
 import {
-  ActionCategoryKey,
+  VestingPaymentsWidgetData,
+  VestingPaymentsWidgetVersion,
   Widget,
+  WidgetId,
   WidgetLocation,
   WidgetVisibilityContext,
 } from '@dao-dao/types'
 
-import { makeManageVestingActionMaker } from './actions/ManageVesting'
-import { VESTING_PAYMENTS_WIDGET_ID } from './constants'
 import { Renderer } from './Renderer'
-import { VestingPaymentsData, VestingPaymentsWidgetVersion } from './types'
 import { VestingPaymentsEditor as Editor } from './VestingPaymentsEditor'
 
-export const VestingPaymentsWidget: Widget<VestingPaymentsData> = {
-  id: VESTING_PAYMENTS_WIDGET_ID,
+export const VestingPaymentsWidget: Widget<VestingPaymentsWidgetData> = {
+  id: WidgetId.VestingPayments,
   Icon: PaidOutlined,
   location: WidgetLocation.Tab,
   visibilityContext: WidgetVisibilityContext.Always,
@@ -24,16 +23,4 @@ export const VestingPaymentsWidget: Widget<VestingPaymentsData> = {
   },
   Renderer,
   Editor,
-  getActionCategoryMakers: (data) => {
-    // Make makers in outer function so they're not remade on every render.
-    const actionMakers = [makeManageVestingActionMaker(data)]
-
-    return [
-      () => ({
-        // Add to Treasury category.
-        key: ActionCategoryKey.Treasury,
-        actionMakers,
-      }),
-    ]
-  },
 }
