+++ packages/stateful/widgets/widgets/WyndDeposit/index.ts
@@ -2,6 +2,7 @@ import {
   ChainId,
   TokenType,
   Widget,
+  WidgetId,
   WidgetLocation,
   WidgetVisibilityContext,
 } from '@dao-dao/types'
@@ -12,7 +13,7 @@ import { WyndDepositEditor as Editor } from './WyndDepositEditor'
 import { WyndDepositRenderer as Renderer } from './WyndDepositRenderer'
 
 export const WyndDepositWidget: Widget<WyndDepositData> = {
-  id: 'wynd_deposit',
+  id: WidgetId.WyndDeposit,
   location: WidgetLocation.Home,
   visibilityContext: WidgetVisibilityContext.Always,
   // WYND only available on Juno mainnet.
