+++ packages/stateless/components/vesting/VestingPaymentCard.stories.tsx
@@ -1,17 +1,16 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
-import { ButtonLink } from '@dao-dao/stateless'
+import { ButtonLink, EntityDisplay } from '@dao-dao/stateful'
 import { makeProps as makeTokenCardProps } from '@dao-dao/stateless/components/token/TokenCard.stories'
 import { CHAIN_ID } from '@dao-dao/storybook'
 import { DaoPageWrapperDecorator } from '@dao-dao/storybook/decorators/DaoPageWrapperDecorator'
 import { EntityType, TokenType } from '@dao-dao/types'
 
-import { EntityDisplay } from '../../../../../components/EntityDisplay'
 import { VestingPaymentCard } from './VestingPaymentCard'
 
 export default {
   title:
-    'DAO DAO / packages / stateful / widgets / widgets / VestingPayments / components / VestingPaymentCard',
+    'DAO DAO / packages / stateless / components / vesting / VestingPaymentCard',
   component: VestingPaymentCard,
   decorators: [DaoPageWrapperDecorator],
 } as ComponentMeta<typeof VestingPaymentCard>
