+++ packages/stateless/components/vesting/index.ts
@@ -1,0 +1,3 @@
+export * from './VestingPaymentCard'
+export * from './VestingPaymentLine'
+export * from './VestingStepsLineGraph'
