+++ packages/types/index.ts
@@ -20,6 +20,7 @@ export * from './stateless'
 export * from './token'
 export * from './units'
 export * from './utils'
+export * from './vesting'
 export * from './voting-module-adapter'
 export * from './wallet'
 export * from './widgets'
