+++ packages/utils/index.ts
@@ -28,3 +28,4 @@ export * from './string'
 export * from './time'
 export * from './token'
 export * from './url'
+export * from './vesting'
